Improve error messaging
